refactor: change containers directory name into pages
