--- packages/client/src/containers/index.ts
@@ -1,6 +1,0 @@
-export { default as Dashboard } from './Dashboard';
-export { default as Login } from './Login';
-export { default as PageNotFound } from './PageNotFound';
-export { default as PasswordRecovery } from './PasswordRecovery';
-export { default as UpdatePassword } from './PasswordRecovery/UpdatePassword';
-export { default as Registration } from './Registration';
