+++ packages/client/src/pages/autorized/Dashboard/Dashboard.styled.ts
