+++ packages/client/src/pages/index.ts
@@ -1,0 +1,6 @@
+export { default as Dashboard } from './autorized/Dashboard';
+export { default as Login } from './unautorized/Login';
+export { default as PageNotFound } from './unautorized/PageNotFound';
+export { default as PasswordRecovery } from './unautorized/PasswordRecovery';
+export { default as UpdatePassword } from './unautorized/PasswordRecovery/UpdatePassword';
+export { default as Registration } from './unautorized/Registration';
