+++ packages/client/src/pages/unautorized/Registration/Registration.styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Button } from '../../elements/buttons';
+import { Button } from '../../../elements/buttons';
 
 export const StyledButton = styled(Button)`
   margin-bottom: 20px;
